Remove old GIN OAuthProvider
This was the old struct + methods that were written to handle DOI
registration in the pre-GOGS GIN. It has been unused code since then and
is now being removed. The only still used function, genSSHKey(), is now
defined in the gogsoap.go (the counterpart file for GOGS).

Minor change: The description (title) of the SSH key that gets added to
the user's profile is now "GIN DOI".

This should be a temporary change since we will probably have the DOI
user handle cloning soon.