Configurable known_hosts file
The service needs the GIN server host key for securely doing the git
clone (with strict host key checking). gindoid now accepts a command
line option for specifying a known_hosts file. The known key for
gin.g-node.org is added here and is copied into the container in the
Dockerfile.