Support multiple email recipients for notification
Instead of setting a single email address on the command line, the
service now accepts a filepath that should contain a list of email
address (one per line). Whenever a request is received, the notification
is sent to all email addresses listed in the file.

The message has also been changed to add the From and To addresses into
the email message header and not just specified through the SMTP
commands. This puts all the recipient addresses in the email To: line,
instead of BCCing the recipients.

The file is read every time a message is about to be sent, so the
recipients can be changed without restarting the service.