Use Client configurations for DataSource
Import the Server configuration structs from the gin-client config
package and use those to store the configs for communicating with the
server.

Read in the DOI user's username and password from the configuration and
use those credentials to clone the repository to be registered.  On
startup, the service checks if the credentials are valid and exits with
an error if they are not.

GinURL and GitURL (previously GinGitURL) are now methods that return the
rendered string.  They call the corresponding AddressStr() methods on
the member struct.