Invalid request handling: General message for user
Telling the user that a specific piece of information is missing from
the URL query string isn't very useful; the user isn't constructing the
string.  Instead, if a user comes across this error, it's likely there
is a misconfiguration with the GIN server or they edited the URL
accidentally.

The new general Invalid Request message is rendered whenever any of the
fields are missing from the query string and the user is instructed to
contact us if they believe they did everything correctly and arrived at
the page (e.g., if the GIN server is misconfigured or something is
broken).

The same message is rendered if the token fails to decrypt.  This can
happen if the key configuration is wrong (if gin-web and gin-doi have
different keys) or if the user somehow edited the encrypted token
string.  The current invalid request message is preferable to a 500.