Small config redesign: Use config directory and file
Full in-memory configuration handling is too troublesome for the current
way that gin-cli works.  There is a plan to redesign a lot of things in
gin-cli, but for now, it's better to just use a configuration file the
same way the client does and set it up on startup with the configuration
keys given to gin-doi.