Rename struct fields to new usage and verify on DO
1. The request data coming from GIN has changed, but we were reusing old
struct field names until now.

2. Since the request data is plugged into a inline function on the Init
page before being submitted to the '/do' route, this could be exploited
to send unauthorised requests by changing the repository path in the
inline script and sending the request.  More straightforward exploit
would be to just send the username and repository payload to the '/do'
route.  Now we verify the request in the DoDOIJob() as well.