Send email early (before starting archive job)
Sending the notification email before beginning the archive job is good
for a couple of reasons:
- We get the notification as soon as the service knows it's valid and it
doesn't have to wait for a long running job to complete.  Archive jobs
can take a very long time if the repository is big (hundreds of GiB).
- If a repository causes issues (extremely large repository, an uncaught
error in the archiving process, etc) that put the service in a faulty
state, one which prevents sending emails, the request isn't lost.

One downside is that if any issues occur during archiving, we no longer
get them in the notification email.  This is currently remedied by
resending the email with error messages attached when needed.