Request result page redesign
startDOIRegistration() function defers anonymous function that sends
email and renders a submission status page.  If the email fails to send,
the user is informed of the failure and asked to contact the admins.
Specific messages and 3 status levels (success, warning, error) can be
assigned to the status page for the user.