[web] Verify request first
Before doing anything else with a request, verify it.  If it's invalid,
just show the error to the user and don't notify the admins.

A new function, renderResult(), renders the results template.  If the
rendering fails, it shows the intended message on a plain page and
returns with an error.