Split funding references to funder and award number
The datacite.yml template requests funding references to separate funder
name from award number with a comma. When rendering the XML file, we can
split on comma to populate the funding field more completely.

If there are no commas, the complete reference is returned only for the
funder name field, like we did before this change.