Display error directly, no Missing slice
Fields with missing values are listed in the error message coming from
the readRepoYAML function, so we can display the message to the user
directly.
Any error that arises is displayed to the user.  The unmarshalling error
has been changed to be more friendly ("error while reading...").
The condition chain is much simpler now.