Reconstruct references from DataCite
New function for building references when rendering pages.  The function
is needed for when we reload data from an XML rather than getting it
from the YAML file.  In the XML, the citations (title and authors) of
the reference are separated from the core info (ID and reltype), so we
need to recombine them.  This isn't required if the data is already
available in the YAML struct.