Deduplicate yml file parsing and validation code
The datacite.yml is read, parsed/unmarshalled and validated at two
points in the registration flow: During request preparation and after
the request is submitted.  A lot of code was duplicated for these two
parts to manage logging and error reporting separately, but it's much
cleaner and easier to work with if we have a single function that
manages both.

The user-friendly error message is returned from the new function and
can be displayed on the page immediately.  More detailed error logging
happens inside the function.