Open issue on XMLRepo in notifyAdmin
Along with sending an email to the admins, open an issue with the same
information and the XML file embedded on the XMLRepo.

Ideally, this would have been a Pull Request on the repository with the
XML file as new content, but opening PRs isn't supported by the API.
Secondarily, it would have been useful to have the XML file as an
attachment to the new issue, but that isn't supported either.