[dataset] Handle prepDir error
Errors returning from the prepDir are now collected and sent
to the admin.
In the long run it might be good to split prepDir to properly
handle resulting errors: if the preparation directory cannot
be created, it does not make sense to continue. If the .htaccess
file cannot be created carrying on with a warning might be ok.